requirements.txt
h5py == 2.9.0
matplotlib == 3.0.3
numpy == 1.14.3
scikit-learn == 0.20.3
scipy == 1.1.0
tensorflow-gpu == 1.3.0
